Food-Stories/requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
Django==2.2.8
django-3-jet==1.0.8
django-allauth==0.42.0
django-ckeditor==5.9.0
django-jet==1.0.8
django-js-asset==1.2.2
django-social-auth==0.7.28
djangorestframework==3.11.0
httplib2==0.18.1
idna==2.10
oauth2==1.9.0.post1
oauthlib==3.1.0
Pillow==7.2.0
psycopg2-binary==2.8.5
python-openid==2.2.5
python3-openid==3.2.0
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
